% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{add_check}
\alias{add_check}
\title{Add a custom check to the validator}
\usage{
add_check(validator, description, condition)
}
\arguments{
\item{validator}{A \code{Validator} object to which the custom check will be added.}

\item{description}{A description of the custom check.}

\item{condition}{Expression to be evaluated or logical conditions to define the custom check. Optional if outcome is set}
}
\value{
The updated \code{Validator} object with the custom check added.
}
\description{
This function allows you to add a custom check to the \code{Validator}object.
}
