% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{add_check_custom}
\alias{add_check_custom}
\title{Add a custom check to the validator}
\usage{
add_check_custom(
  validator,
  description,
  outcome,
  type = c("error", "warning", "note")
)
}
\arguments{
\item{validator}{A \code{Validator} object to which the custom check will be added.}

\item{description}{A description of the custom check.}

\item{outcome}{Logical vector indicating the result of the check (TRUE/FALSE).
Outcome must be logical.}

\item{type}{The type of the check, can be "error", "warning", or "note".}
}
\value{
The updated \code{Validator} object with the custom check added.
}
\description{
This function allows you to add a custom check outcomes to the validator log. The outcomes must be a logical vector.
}
