% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{check}
\alias{check}
\title{Validate a Validator Object}
\usage{
check(validator, ...)
}
\arguments{
\item{validator}{An object of class \code{Validator} to be validated.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The validated \code{Validator} object if all checks pass. If any check fails, an error is thrown.
}
\description{
This function runs the full suite of validation checks on a \code{Validator} object.
}
\examples{
# create schema
schema <- list(
  check_duplicates = FALSE,
  check_completeness = FALSE,
  columns = list(
    age = list(type = "double", optional = FALSE),
    sex = list(type = "character", optional = FALSE)
  )
)

# create dataframe
df <- data.frame(
  age = c(10, 11, 13, 15, 22, 34, 80),
  sex = c("M", "F", "M", "F", "M", "F", "M")
)

# create validator object
validator <- new_validator(
  data = df,
  schema = schema
)
# validate the data
validator <- check(validator)

}
