% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_backseries}
\alias{check_backseries}
\title{Check backseries consistency}
\usage{
check_backseries(validator)
}
\arguments{
\item{validator}{A Validator object containing the schema and agent.}
}
\value{
The updated Validator object with outcomes logged.
}
\description{
Checks if the latest data is consistent with previous data.
}
