% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_colnames}
\alias{check_colnames}
\title{Check Column Names against schema}
\usage{
check_colnames(validator)
}
\arguments{
\item{validator}{A \code{Validator} object containing the column names to be checked.}
}
\value{
The updated \code{Validator} object with QA entries for each check.
}
\description{
This function performs checks on the column names of a \code{Validator} object to ensure they follow specific naming conventions and
meet schema conditions.
}
\details{
The function performs the following checks on the column names:
\itemize{
\item Ensures column names do not contain spaces.
\item Ensures column names do not contain symbols other than underscores.
\item Ensures column names do not contain uppercase letters.
}

For each check, a QA entry is added to the \code{Validator} object with details about the check, whether it passed, and the IDs of failing columns (if any). # nolint: line_length_linter.
}
