% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_column_contents}
\alias{check_column_contents}
\title{Check Column Contents against schema and checks}
\usage{
check_column_contents(validator)
}
\arguments{
\item{validator}{A \code{Validator} object containing the column names to be checked.}
}
\value{
The updated \code{Validator} object with QA entries for each check.
}
\description{
This function performs checks on the columns of \code{Validator$data} to ensure they meet the
specified schema conditions and checks.
}
