% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_completeness}
\alias{check_completeness}
\title{Check dataset for missing columns}
\usage{
check_completeness(validator)
}
\arguments{
\item{validator}{data \code{Validator} object}
}
\value{
The updated validator object with new log entries appended.
}
\description{
Check dataset for missing columns
}
