% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_schema_contents_against_df}
\alias{check_schema_contents_against_df}
\title{Check schema contents against the data frame provided}
\usage{
check_schema_contents_against_df(validator)
}
\arguments{
\item{validator}{A \code{Validator} object containing the data and schema to check against.}
}
\value{
The updated \code{Validator} object with QA entries added for any issues found in the schema.
}
\description{
This function checks that the contents of the schema are consistent with the data frame provided.
It checks for unused schema entries, incompatible schema entries, and that any columns specified
in the schema are present in the data frame.
}
