% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{date_from_schema}
\alias{date_from_schema}
\title{Attempt to parse a date from the schema and capture any warnings
This function tries to parse a date string using \code{lubridate::ymd} and captures any warnings that occur during parsing.
If a warning is raised, the warning message is returned; otherwise, \code{NULL} is returned.}
\usage{
date_from_schema(date_value)
}
\arguments{
\item{date_value}{A date string to be parsed.}
}
\value{
A warning message if parsing fails, otherwise \code{NULL}.
}
\description{
Attempt to parse a date from the schema and capture any warnings
This function tries to parse a date string using \code{lubridate::ymd} and captures any warnings that occur during parsing.
If a warning is raised, the warning message is returned; otherwise, \code{NULL} is returned.
}
