% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{decimal_places}
\alias{decimal_places}
\title{Check Decimal Places in Numeric Columns}
\usage{
decimal_places(x, column_name)
}
\arguments{
\item{x}{A numeric vector.}

\item{column_name}{A string passed from a DataFrame}
}
\value{
A vector of the same length as \code{x}, indicating the number of decimal
places for each element. If an element is \code{NA}, it returns \code{NA}.
}
\description{
This function calculates the number of decimal places in a numeric vector.
}
