% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{hard_checks_status}
\alias{hard_checks_status}
\title{Check the status of errors and warnings in the validator log}
\usage{
hard_checks_status(validator, hard_check)
}
\arguments{
\item{validator}{A \code{Validator} object to check the log.}

\item{hard_check}{A logical value indicating whether to perform hard checks (default is TRUE).}
}
\value{
Warning if there are any warnings or errors in the log when \code{hard_check} is FALSE.
Error if there are any errors and \code{hard_check} is TRUE.
}
\description{
This function raises errors or warnings if any checks flagged as error or warnings fail.
}
