% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{iqr_bounds}
\alias{iqr_bounds}
\title{Flag outliers based on Interquartile Range (IQR).
Outliers are flagged if they are below Q1 - (mulitplier * IQR) or above Q3 + (multiplier * IQR).}
\usage{
iqr_bounds(x, multiplier = 1.5)
}
\arguments{
\item{x}{A numeric vector.}

\item{multiplier}{A numeric value to multiply the IQR by (default is 1.5).}
}
\value{
A vector the same size as \code{x}, with \code{TRUE} for values that are outliers and \code{FALSE} otherwise
}
\description{
Flag outliers based on Interquartile Range (IQR).
Outliers are flagged if they are below Q1 - (mulitplier * IQR) or above Q3 + (multiplier * IQR).
}
