% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{is_column_contents_valid}
\alias{is_column_contents_valid}
\title{Check column contents valid}
\usage{
is_column_contents_valid(schema)
}
\arguments{
\item{schema}{the validator schema}
}
\value{
\code{TRUE} if all column values are valid, otherwise an error is raised.
}
\description{
This wrapper calls is_valid_column_values for each column in the schema
}
