% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{is_type_valid}
\alias{is_type_valid}
\title{Check type of column in schema is valid}
\usage{
is_type_valid(schema)
}
\arguments{
\item{schema}{the validator schema}
}
\value{
\code{TRUE} if all column types are valid, otherwise an error is raised.
}
\description{
Check type of column in schema is valid
}
