% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{is_valid_column_values}
\alias{is_valid_column_values}
\title{Check that max values are not less than min values in column schema}
\usage{
is_valid_column_values(column_schema, col_name)
}
\arguments{
\item{column_schema}{A list representing the schema for a specific column, which may contain max and min value specifications.}

\item{col_name}{The name of the column being checked, used for error messages.}
}
\value{
\code{TRUE} if all max values are greater than or equal to their corresponding min values, otherwise an error is raised.
}
\description{
This function checks that for any column schema, the max values
(e.g., max_string_length, max_date) are not less than the corresponding min values
(e.g., min_string_length, min_date). If any such inconsistency is found, an error
is raised with a descriptive message.
}
