% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{is_valid_schema}
\alias{is_valid_schema}
\title{Check if the schema is valid}
\usage{
is_valid_schema(schema)
}
\arguments{
\item{schema}{A list to validate.}
}
\value{
\code{TRUE} if the schema is a valid named list, otherwise \code{FALSE}.
}
\description{
Check if the schema is valid
}
