% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{log_html}
\alias{log_html}
\title{Generate HTML Representation of a Log}
\usage{
log_html(log)
}
\arguments{
\item{log}{A log object to be converted into HTML. It is expected to be in a format compatible with \code{log_to_table}.}
}
\value{
A string containing the HTML representation of the log.
}
\description{
Generate HTML Representation of a Log
}
