% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{log_to_table}
\alias{log_to_table}
\title{Convert Validator Log to Table}
\usage{
log_to_table(log)
}
\arguments{
\item{log}{A list representing the validator log, where each element is a log entry.}
}
\value{
A data frame containing the formatted log entries.
}
\description{
This function converts a validator log into a formatted data frame (table) for exports.
}
