% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{print.Validator}
\alias{print.Validator}
\title{Print Validator Log}
\usage{
\method{print}{Validator}(validator)
}
\arguments{
\item{validator}{A \code{Validator} object containing a log to be printed.}
}
\value{
A markdown-formatted table of the validator log.
}
\description{
This function prints the log of a \code{Validator} object in a markdown table format.
}
