% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{run_checks}
\alias{run_checks}
\title{Run column checks}
\usage{
run_checks(validator, i)
}
\arguments{
\item{validator}{\code{Validator} object passed from check_column_contents.}

\item{i}{column index}
}
\value{
validator object
}
\description{
To be used by check_column_contents - not intended to be run separately.
}
