% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{types_to_classes}
\alias{types_to_classes}
\title{Convert complex types to the correct types and classes}
\usage{
types_to_classes(schema)
}
\arguments{
\item{schema}{A list containing a \code{columns} element, where each column is a list with a \code{type} field.}
}
\value{
The modified schema with updated \code{type} and \code{class} fields for each column.
}
\description{
This function modifies a schema by converting column types to their corresponding R classes.
}
