% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validate_and_convert_date_formats}
\alias{validate_and_convert_date_formats}
\title{Validate date formats in the schema
This function checks that any date formats specified in the schema are valid and can be parsed correctly.}
\usage{
validate_and_convert_date_formats(schema)
}
\arguments{
\item{schema}{A list containing a \code{columns} element, where each column may have \code{min_date} and \code{max_date} fields.}
}
\value{
The original schema if all date formats are valid. If any date format is invalid, an error is thrown with a message indicating the issue.
}
\description{
Validate date formats in the schema
This function checks that any date formats specified in the schema are valid and can be parsed correctly.
}
