% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{z_score}
\alias{z_score}
\title{Check Z Score of Numeric Columns}
\usage{
z_score(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A vector of the same length as \code{x}, indicating the z-score for each element.
}
\description{
This function calculates the maximum z-score for a numeric column.
}
