% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{check_and_export}
\alias{check_and_export}
\title{Validate data against a schema and output results}
\usage{
check_and_export(
  data,
  schema,
  file,
  format,
  hard_check = FALSE,
  backseries = NULL,
  name = deparse(substitute(data))
)
}
\arguments{
\item{data}{The data to be validated.}

\item{schema}{The schema to validate against.}

\item{file}{The file path where the validation results will be exported.}

\item{format}{The format in which the validation results will be exported.}

\item{hard_check}{logical. Optional - FALSE by default. If TRUE, raises an error if there are any failed checks. Otherwise, raises a warning.}

\item{backseries}{A previous version of the data to check against (optional).}

\item{name}{validator name - defaults to the name of the dataframe object supplied to "data" (Optional).
Must be a single character string.}
}
\value{
The exported validation results.
}
\description{
This function validates data against a given schema, performs checks,
and exports the validation results to a specified file in a given format.
}
