% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check for duplicate rows}
\usage{
check_duplicates(validator)
}
\arguments{
\item{validator}{\code{Validator} object}
}
\description{
Check for duplicate rows
}
