% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.checker-package.R
\docType{package}
\name{data.checker-package}
\alias{data.checker}
\alias{data.checker-package}
\title{data.checker: Data Checker for Validating Data Frames Against Defined Schema}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Validates data frames against a defined schema. Produces a report of the checks performed and any issues found, with index and entry value where appropriate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://onsdigital.github.io/data.checker/}
}

}
\author{
\strong{Maintainer}: Analysis Standards and Pipelines Team (ONS) \email{ASAP@ons.gov.uk}

Other contributors:
\itemize{
  \item Crown Copyright [copyright holder]
}

}
\keyword{internal}
