% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{export}
\alias{export}
\title{Generic export function}
\usage{
export(object, ...)
}
\arguments{
\item{object}{The object to be checked.}

\item{...}{Additional arguments passed to specific methods.}
}
\description{
This function exists For ease of use - see export.Validator() for details.
}
