% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export.Validator}
\alias{export.Validator}
\title{Export Validator Log}
\usage{
\method{export}{Validator}(object, file, format = c("yaml", "json", "html", "csv"), ...)
}
\arguments{
\item{object}{A \code{Validator} object containing the log to be exported.}

\item{file}{A string specifying the file path where the log will be exported.
The file extension must match the specified format.}

\item{format}{A string specifying the format of the output file.
Supported formats are \code{"yaml"}, \code{"json"}, \code{"html"}, and \code{"csv"}.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Writes the log to the specified file. No value is returned.
}
\description{
This function exports the log of a \code{Validator} object to a file in the specified format.
}
