% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{new_validator}
\alias{new_validator}
\title{Validator Constructor}
\usage{
new_validator(
  data,
  schema,
  backseries = NULL,
  name = deparse(substitute(data))
)
}
\arguments{
\item{data}{A data frame to validate against the schema.}

\item{schema}{A schema object that defines the validation rules. See the vignette for more details on schema structure.
This can also be a file path to a JSON, YAML, or TOML file containing the schema.}

\item{backseries}{A previous version of the data to check against (optional).}

\item{name}{validator name - defaults to the name of the dataframe object supplied to "data" (optional).
Must be a single character string.}
}
\value{
An object of class \code{Validator}.
}
\description{
Creates a \code{Validator} object to validate data against a given schema.
}
\examples{
# create schema
schema <- list(
  check_duplicates = FALSE,
  check_completeness = FALSE,
  columns = list(
    age = list(type = "double", optional = FALSE),
    sex = list(type = "character", optional = FALSE)
  )
)

# create dataframe
df <- data.frame(
  age = c(10, 11, 13, 15, 22, 34, 80),
  sex = c("M", "F", "M", "F", "M", "F", "M")
)

# create validator object
validator <- new_validator(
  data = df,
  schema = schema
)
}
