% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{new_validator}
\alias{new_validator}
\title{Validator Constructor}
\usage{
new_validator(data, schema, backseries = NULL, name = NULL)
}
\arguments{
\item{data}{A data frame to validate against the schema.}

\item{schema}{A schema object that defines the validation rules.}

\item{backseries}{A previous version of the data to check against (optional).}

\item{name}{Optional validator name. Must be a single character string.}
}
\value{
An object of class \code{Validator}.
}
\description{
Creates a \code{Validator} object to validate data against a given schema.
}
