% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{new_validator}
\alias{new_validator}
\title{Validator Constructor}
\usage{
new_validator(data, schema)
}
\arguments{
\item{data}{A data frame to validate against the schema.}

\item{schema}{A schema object that defines the validation rules.}
}
\value{
An object of class \code{Validator}.
}
\description{
Creates a \code{Validator} object to validate data against a given schema.
}
